[Silver I] Title: 더 게임 오브 데스, Time: 104 ms, Memory: 14124 KB -BaekjoonHub

+++ 백준/Silver/32248. 더 게임 오브 데스/더 게임 오브 데스.java
@@ -54,7 +54,7 @@ public class Main {
          
 
         T=T-start;
-        T=(long)T%cyclesize ;
+        T= T%cyclesize ;
         
       // System.out.println(  cyclesize+ " "+start  );
       // System.out.println(  T+start  );
